// frontend/src/generated/janus_common.ts
export enum Protocol {
  PROTOCOL_UNSPECIFIED = 0,
  PROTOCOL_ANY = 1,
  PROTOCOL_TCP = 2,
  PROTOCOL_UDP = 3,
  PROTOCOL_HTTP = 4,
  PROTOCOL_OTHER = 5,
  UNRECOGNIZED = -1,
}
// ...
export enum Verdict {
  VERDICT_UNSPECIFIED = 0,
  VERDICT_ALLOW = 1,
  VERDICT_DROP = 2,
  UNRECOGNIZED = -1,
}
// ...
export enum EngineStage {
  ENGINE_STAGE_UNSPECIFIED = 0,
  ENGINE_STAGE_PREPROCESS = 1,
  ENGINE_STAGE_POLICY = 2,
  ENGINE_STAGE_VECTOR_FILTER = 3,
  ENGINE_STAGE_REGEX = 4,
  ENGINE_STAGE_AHO = 5,
  ENGINE_STAGE_DPI = 6,
  ENGINE_STAGE_ENCAPSULATION = 7,
  ENGINE_STAGE_TRANSMISSION = 8,
  ENGINE_STAGE_LOGGING = 9,
  UNRECOGNIZED = -1,
}